update shared utility

+++ mascot/cvFunction.cpp
@@ -21,8 +21,8 @@
 #include "../svm-shared/smoSolver.h"
 #include "../svm-shared/Cache/cache.h"
 #include "../svm-shared/fileOps.h"
-#include "../DataReader/BaseLibsvmReader.h"
-#include "../DataReader/LibsvmReaderSparse.h"
+#include "../SharedUtility/DataReader/BaseLibsvmReader.h"
+#include "../SharedUtility/DataReader/LibsvmReaderSparse.h"
 
 using std::cout;
 using std::endl;
@@ -36,7 +36,7 @@ void gridSearch(Grid &SGrid, string strTrainingFileName){
 
 	int nNumofFeature = 0;
 	int nNumofInstance = 0;
-	long long nNumofValue = 0;
+	uint nNumofValue = 0;
 	BaseLibSVMReader::GetDataInfo(strTrainingFileName, nNumofFeature, nNumofInstance, nNumofValue);
 	LibSVMDataReader drHelper;
 	drHelper.ReadLibSVMAsDense(v_vDocVector, v_nLabel, strTrainingFileName, nNumofFeature);
